+++ CHANGELOG.md
@@ -53,11 +53,13 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Improvements
 
+- (erc20) [\#677](https://github.com/tharsis/evmos/pull/677) Add Amino registration to `ConvertCoin` and `ConvertERC20` msgs for ERC712 compatibility.
 - (deps) [\#668](https://github.com/tharsis/evmos/pull/668) Bump Cosmos SDK to [`v0.45.5`](https://github.com/cosmos/cosmos-sdk/releases/tag/v0.45.5)
 - (erc20) [\#642](https://github.com/tharsis/evmos/pull/642) Remove enforcing ibc and channel names during `RegisterCoin`
 
 ### Bug Fixes
 
+- (app) [\#682](https://github.com/tharsis/evmos/pull/682) Fix Tendermint consensus params (Evidence `MaxAgeNumBlocks` and `MaxAgeDuration`)
 - (incentives) [\#656](https://github.com/tharsis/evmos/pull/656) Fix incentives that were previously only allocated to `EthAccount`s.
 - (fees) [\#612](https://github.com/tharsis/evmos/pull/612) Fix fees registration cli command and description
 - (inflation) [\#554](https://github.com/tharsis/evmos/pull/554) Changing erroneous epoch skips to `daily` instead of `weekly`